added validation of meals collection by user's id

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -3,11 +3,14 @@ package ru.javawebinar.topjava.repository.mock;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
@@ -21,6 +24,7 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
     public Meal save(Meal meal) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
+            meal.setUserId(SecurityUtil.authUserId());
             repository.put(meal.getId(), meal);
             return meal;
         }
@@ -30,17 +34,23 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public void delete(int id) {
-        repository.remove(id);
+        Meal currentMeal = repository.get(id);
+        if (currentMeal.getUserId() == SecurityUtil.authUserId()) {
+            repository.remove(id);
+        }
     }
 
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal currentMeal = repository.get(id);
+        return currentMeal.getUserId() == SecurityUtil.authUserId() ? currentMeal : null;
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        return repository.values().stream().filter(meal -> meal.getUserId() == SecurityUtil.authUserId())
+                .sorted(Comparator.comparing(Meal::getDate))
+                .collect(Collectors.toList());
     }
 }
